python-api/requirements.txt
anthropic
fastapi
# pymongo[srv]
uvicorn
python-dotenv

pymongo
motor